# startup/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Full documentation of App APIs",
    "version": "1.0.0",
    "title": "App APIs",
    "contact": {
      "email": "App@Developer.com"
    }
  },
...
  "paths": {
...
    "/users/register": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "user's parameters",
            "description": "user email, Password and phone number. ",
            "schema": {
              "$ref": "#/definitions/userModel"
            }
          }
        ],
        "tags": [
          "User"
        ],
        "summary": "Creates a user",
        "operationId": "createUser",
        "description": "accepts user parameters then creates a user by storing it into database.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "email already exists or missing parameters"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "userModel": {
      "type": "object",
      "required": [
        "email",
        "password",
        "name",
        "phone",
        "role"
      ],
      "properties": {
        "email": {
          "type": "string",
          "example": "example@example.com"
        },
        "password": {
          "type": "string",
          "example": "12345678"
        },
        "name": {
          "type": "string",
          "example": "Alaa thrwat"
        },
        "phone": {
          "type": "string",
          "example": "01234567890"
        },
        "role": {
          "type": "string",
          "example": "admin"
        }
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "host": "localhost:7500",
  "basePath": "/api"
}